Relocate existing Account class to model_old.
