+++ common/src/main/java/com/microsoft/identity/common/internal/cache/IShareSingleSignOnState.java
@@ -1,6 +1,6 @@
 package com.microsoft.identity.common.internal.cache;
 
-import com.microsoft.identity.common.Account;
+import com.microsoft.identity.common.model_old.Account;
 import com.microsoft.identity.common.internal.providers.oauth2.RefreshToken;
 
 /**
